fixed minor and major bugs (mostly related to the hex file being too large (>8k)!

+++ Project/Latencytest_STM8/main.h
@@ -1,6 +1,6 @@
 #ifndef __MAIN_H
 #define __MAIN_H
 
-uint16_t time_now(void);
+uint32_t time_now(void);
 
 #endif
